+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added:
 - Added required levelz lvl compat
+- Added per difficulty required items
 ### Fixed:
 - 
 ### Changed: